cpu load on metrics overview

+++ api/src/main/java/com/cloud/server/ManagementServerHostStats.java
@@ -103,7 +103,7 @@ public interface ManagementServerHostStats {
 
     double[] getSystemLoadAverages();
 
-    double[] getSystemCyclesUsage();
+    long[] getSystemCyclesUsage();
 
     boolean isDbLocal();
